fixup! Add users and hobbies.

+++ src/hobbies/hobbies.controller.ts
@@ -6,6 +6,8 @@ import {
   Patch,
   Param,
   Delete,
+  HttpException,
+  HttpStatus,
 } from '@nestjs/common';
 import { HobbiesService } from './hobbies.service';
 import { CreateHobbyDto } from './dto/create-hobby.dto';
@@ -40,7 +42,13 @@ export class HobbiesController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const { hobbies } = await this.findOne(id);
+
+    if (!hobbies[0]) {
+      throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+    }
+
     return this.hobbiesService.remove(+id, loggedInUserId);
   }
 }
